% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{console-methods}
\alias{console-methods}
\alias{char}
\alias{stat}
\alias{func}
\alias{cat_line}
\title{Methods for printing to console}
\usage{
char(x)

stat(...)

func(x)

cat_line(...)
}
\arguments{
\item{x}{Character}

\item{...}{Objects to print}
}
\description{
Print to console using colours.
}
