% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{docker_ids_get}
\alias{docker_ids_get}
\title{Get docker names for a module}
\usage{
docker_ids_get(pkgnm)
}
\arguments{
\item{pkgnm}{Package name of module}
}
\value{
Logical
}
\description{
Return the image and container names for a module. Will attempt
to build/pull image if missing.
}
\seealso{
Other ids: 
\code{\link{meta_get}()},
\code{\link{modules_list}()}
}
\concept{ids}
