% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_img_ls}
\alias{docker_img_ls}
\title{List the number of installed images}
\usage{
docker_img_ls()
}
\value{
tibble
}
\description{
Return a table of all the available Docker images.
}
\concept{docker}
