% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{guess_pkgnm}
\alias{guess_pkgnm}
\title{Guess package name}
\usage{
guess_pkgnm(repo)
}
\arguments{
\item{repo}{Repository (e.g. GitHub) associated with module}
}
\value{
character(1)
}
\description{
Return package name from a repo name.
}
\details{
Raises error if no module discovered.
}
\seealso{
Other ids: \code{\link{docker_ids_get}},
  \code{\link{img_get}}, \code{\link{meta_get}},
  \code{\link{modules_list}}
}
\concept{ids}
