% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{image_install}
\alias{image_install}
\title{Install module's image}
\usage{
image_install(pkgnm, tag = "latest", pull = TRUE)
}
\arguments{
\item{pkgnm}{Name of module's R package}

\item{tag}{Docker tag, default 'latest'}

\item{pull}{Pull from Docker Hub or build locally? Default, FALSE.}
}
\value{
Integer
}
\description{
Install the Docker image for an outsider module after the module
package has been installed.
}
