% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{img_get}
\alias{img_get}
\title{Construct Docker image name from meta}
\usage{
img_get(pkgnm)
}
\arguments{
\item{pkgnm}{Package name of module}
}
\value{
character(1)
}
\description{
Return Docker image name using DockerHub username, if available.
}
\seealso{
Other ids: \code{\link{docker_ids_get}},
  \code{\link{guess_pkgnm}}, \code{\link{meta_get}},
  \code{\link{modules_list}}
}
\concept{ids}
