% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install module}
\usage{
install(flpth, tag = "latest", pull = FALSE)
}
\arguments{
\item{flpth}{File path to module directory.}

\item{tag}{Docker tag, default 'latest'}
}
\value{
logical
}
\description{
Install outsider module: install R package, build/pull Docker
image.
}
