% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{is_docker_available}
\alias{is_docker_available}
\title{Check if Docker is installed and running}
\usage{
is_docker_available(call_error = TRUE)
}
\arguments{
\item{call_error}{Call an error if no Docker detected? Default TRUE.}
}
\description{
Raises an error if docker is not available.
}
