% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{is_filepath}
\alias{is_filepath}
\title{Is a filepath?}
\usage{
is_filepath(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Logical
}
\description{
Return TRUE or FALSE for whether character(s) is a valid
filepath.
}
\seealso{
Other private: 
\code{\link{log_get}()},
\code{\link{to_basename}()}
}
\concept{private}
