% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{is_installed}
\alias{is_installed}
\title{Is module installed?}
\usage{
is_installed(pkgnm)
}
\arguments{
\item{pkgnm}{Package name}
}
\value{
logical(1)
}
\description{
Return TRUE if module is installed.
}
