% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{is_server_connected}
\alias{is_server_connected}
\title{Is server connected?}
\usage{
is_server_connected()
}
\value{
logical
}
\description{
Return TRUE if an \code{ssh} session exists with which
\code{outsider} can interact.
}
\seealso{
Other private-server: 
\code{\link{server_download}()},
\code{\link{server_fetch}()},
\code{\link{server_upload}()}
}
\concept{private-server}
