% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{meta_get}
\alias{meta_get}
\title{Get outsider module details}
\usage{
meta_get(pkgnm)
}
\arguments{
\item{pkgnm}{Package name of module}
}
\value{
Named list
}
\description{
Return a named list of all metadata associated with a module
}
\seealso{
Other ids: 
\code{\link{docker_ids_get}()},
\code{\link{modules_list}()}
}
\concept{ids}
