% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outsider.R
\docType{package}
\name{outsider.base}
\alias{outsider.base}
\title{outsider.base: Base Package for outsider.}
\description{
For more information visit the outsider website
(\url{https://docs.ropensci.org/outsider/}).
}
