% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{pkg_install}
\alias{pkg_install}
\title{Install outsider module package}
\usage{
pkg_install(flpth, verbose = TRUE)
}
\arguments{
\item{flpth}{File path to module directory.}

\item{verbose}{Be verbose? Default TRUE.}
}
\value{
Logical(1)
}
\description{
Install outsider module's package.
}
