% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{server_connect}
\alias{server_connect}
\title{Connect to a server}
\usage{
server_connect(session)
}
\value{
logical
}
\description{
Connect to a server, make accessible to \code{outsider} and
set-up for \code{outsider} interaction.
}
\seealso{
Other private-server: \code{\link{is_server_connected}},
  \code{\link{server_disconnect}},
  \code{\link{server_download}},
  \code{\link{server_fetch}}, \code{\link{server_upload}}
}
\concept{private-server}
