% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{server_disconnect}
\alias{server_disconnect}
\title{Disconnect from a server}
\usage{
server_disconnect()
}
\value{
logical
}
\description{
Disconnect from a server and remove from \code{outsider}
}
\seealso{
Other private-server: \code{\link{is_server_connected}},
  \code{\link{server_connect}},
  \code{\link{server_download}},
  \code{\link{server_fetch}}, \code{\link{server_upload}}
}
\concept{private-server}
