% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{server_download}
\alias{server_download}
\title{Download from server}
\usage{
server_download(origin, dest)
}
\arguments{
\item{origin}{Origin filepath}

\item{dest}{Destination filepath}
}
\value{
Logical
}
\description{
Download file/folder from connected server. File is copied to
a temporary folder before transferred to desired destination.
}
\seealso{
Other private-server: 
\code{\link{is_server_connected}()},
\code{\link{server_fetch}()},
\code{\link{server_upload}()}
}
\concept{private-server}
