% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{server_fetch}
\alias{server_fetch}
\title{Fetch server "session"}
\usage{
server_fetch(verbose)
}
\arguments{
\item{verbose}{Be verbose? Logical.}
}
\value{
ssh session
}
\description{
Return connected session to server.
}
\details{
See \code{\link[ssh]{ssh_connect}} for more details.
}
\seealso{
Other private-server: 
\code{\link{is_server_connected}()},
\code{\link{server_download}()},
\code{\link{server_upload}()}
}
\concept{private-server}
