% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{server_upload}
\alias{server_upload}
\title{Upload to server}
\usage{
server_upload(fl)
}
\arguments{
\item{fl}{File/folder to be transferred.}
}
\value{
Logical
}
\description{
Upload file/folder to connected server. File is placed in
working dir on server.
}
\seealso{
Other private-server: 
\code{\link{is_server_connected}()},
\code{\link{server_download}()},
\code{\link{server_fetch}()}
}
\concept{private-server}
