% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{uninstall}
\alias{uninstall}
\title{Uninstall and remove a module}
\usage{
uninstall(pkgnm)
}
\arguments{
\item{pkgnm}{Package name}
}
\value{
Logical(1)
}
\description{
Remove outsider module: uninstall package, delete Docker image.
}
\details{
If program is successfully removed TRUE is returned, else FALSE.
}
