% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys.R
\name{exec_wait}
\alias{exec_wait}
\title{Execute system commands and wait for response}
\usage{
exec_wait(
  cmd,
  args = NULL,
  std_out = stdout(),
  std_err = stderr(),
  std_in = NULL,
  timeout = 0
)
}
\arguments{
\item{cmd}{Command}

\item{args}{Arguments}

\item{std_out}{Standard out}

\item{std_err}{Standard error}

\item{std_in}{Standard in}

\item{timeout}{Timeout}
}
\value{
logical
}
\description{
Passes arguments to \code{\link[sys]{exec_wait}}, if a server
is connected arguments are passed to \code{\link[ssh]{ssh_exec_wait}}
instead.
}
\seealso{
Other private-sys: 
\code{\link{exec_internal}()}
}
\concept{private-sys}
