% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install module}
\usage{
install(flpth, tag = "latest", pull = FALSE, verbose = TRUE)
}
\arguments{
\item{flpth}{File path to module directory.}

\item{tag}{Docker tag, default 'latest'}

\item{pull}{Pull from Docker Hub or build locally? Default, FALSE.}

\item{verbose}{Be verbose? Default TRUE.}
}
\value{
Integer
}
\description{
Install outsider module: install R package, build/pull Docker
image. (Will only pull if image is available via DockerHub)
Returns 0 if all successful, 1 if only R package installs, 2 if R package
and Docker image fail.
}
