% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{modules_list}
\alias{modules_list}
\title{List all installed outsider modules}
\usage{
modules_list()
}
\value{
Logical
}
\description{
Return the R package names of all installed outsider modules
}
\seealso{
Other ids: 
\code{\link{docker_ids_get}()},
\code{\link{meta_get}()}
}
\concept{ids}
