% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{server_disconnect}
\alias{server_disconnect}
\title{Disconnect from a server}
\usage{
server_disconnect()
}
\value{
logical
}
\description{
Disconnect from a server and remove from \code{outsider}
}
\details{
This requires installation of \code{ssh} package.
}
\examples{
library(outsider.base)

# NOT RUN
\dontrun{
if (requireNamespace("ssh", quietly = TRUE)) {
  session <- ssh::ssh_connect(host = '[INSERT HOST IP]')
  server_connect(session = session)
  # run outsider.base commands, when finished
  server_disconnect()
}
}
}
\seealso{
Other public-server: 
\code{\link{server_connect}()}
}
\concept{public-server}
